requirements.txt
pillow=="8.1.2"